+++ plutus-pretty/test/properties/Main.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE TemplateHaskell #-}
+{-# OPTIONS_GHC -fno-specialize #-}
 
 module Main (main) where
 
@@ -21,6 +22,7 @@ main =
     [ testProperty "BuiltinString pretty-print" stringSkeletalProp
     , testProperty "Tuple pretty-print" tupleSkeletalProp
     , testProperty "Record pretty-print" recordSkeletalProp
+    , testProperty "Tuple3 pretty-print" tuple3SkeletalProp
     ]
   where
     go :: QuickCheckTests
@@ -42,25 +44,42 @@ tupleSkeletalProp = forAllShrinkShow arbitrary shrink ppShow go
     go :: (BuiltinString, Integer) -> Property
     go tup@(s, i) =
       let lhs =
-            "{ \"fst\": "
+            "{\"fst\": "
               PTx.<> showSkeletal s
               PTx.<> ", \"snd\": "
               PTx.<> showSkeletal i
-              PTx.<> " }"
+              PTx.<> "}"
           rhs = showSkeletal tup
        in lhs === rhs
 
+tuple3SkeletalProp :: Property
+tuple3SkeletalProp = forAllShrinkShow arbitrary shrink ppShow go
+  where
+    go :: (BuiltinString, Integer, BuiltinString) -> Property
+    go tup@(s, i, s') =
+      let lhs =
+            "{\"fst\": "
+              PTx.<> showSkeletal s
+              PTx.<> ", \"snd\": "
+              PTx.<> showSkeletal i
+              PTx.<> ", \"thd\": "
+              PTx.<> showSkeletal s'
+              PTx.<> "}"
+          rhs = showSkeletal tup
+       in lhs === rhs
+
+-- "{ \"recordTag\": \"Foo\", \"fields\": { \"bar\": \"\", \"baz\": 0 } }"
 recordSkeletalProp :: Property
 recordSkeletalProp = forAllShrinkShow arbitrary shrink ppShow go
   where
     go :: Foo -> Property
     go f@(Foo x y) =
       let lhs =
-            "{ \"recordTag\": \"Foo\", \"fields\": { \"bar\": "
+            "{\"recordTag\": \"Foo\", \"fields\": {\"bar\": "
               PTx.<> showSkeletal x
-              PTx.<> " , \"baz\": "
+              PTx.<> ", \"baz\": "
               PTx.<> showSkeletal y
-              PTx.<> " ,  } }"
+              PTx.<> "}}"
        in lhs === showSkeletal f
 
 data Foo = Foo